remove cyclic dependencies between packages by moving methods
